device/requirements.txt
PySide6==6.7.2
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
tenacity==8.4.2
qrcode==7.4.2
pillow==10.4.0
orjson==3.10.7
fastjsonschema==2.20.0
appdirs==1.4.4
loguru==0.7.2
typing_extensions==4.12.2
watchdog==4.0.2
sqlite-utils==3.36
psutil==6.0.0